complement model refactor
