+++ pyzoo/zoo/zouwu/model/forecast/model/model_builder.py
@@ -45,11 +45,17 @@ class ModelBuilder:
             return mapping[class_name]
 
         if dev_option == 'pytorch':
-            from zoo.zouwu.model.forecast.model import PytorchBaseModel
+            from zoo.zouwu.model.forecast.model.base_pytorch_model import PytorchBaseModel
+            from zoo.zouwu.model.forecast.model.Seq2Seq_pytorch import Seq2SeqPytorch
+            from zoo.zouwu.model.forecast.model.VanillaLSTM_pytorch import VanillaLSTMPytorch
+            from zoo.zouwu.model.forecast.model.tcn import TCNPytorch
             return cls(cls=get_class(PytorchBaseModel))
 
         elif dev_option == 'tf.keras':
-            from zoo.zouwu.model.forecast.model import KerasBaseModel
+            from zoo.zouwu.model.forecast.model.base_keras_model import KerasBaseModel
+            from zoo.zouwu.model.forecast.model.VanillaLSTM import VanillaLSTM
+            from zoo.zouwu.model.forecast.model.MTNet_keras import MTNetKeras
+            from zoo.zouwu.model.forecast.model.Seq2Seq import LSTMSeq2Seq
             return cls(cls=get_class(KerasBaseModel))
 
     @classmethod
@@ -59,13 +65,13 @@ class ModelBuilder:
     def build_from_ckpt(self, checkpoint_filename):
         '''Restore from a saved model'''
         if self.backend == "pytorch":
-            from zoo.zouwu.model.forecast.model import PytorchBaseModel
+            from zoo.zouwu.model.forecast.model.base_pytorch_model import PytorchBaseModel
             model = PytorchBaseModel(**self.params)
             model.restore(checkpoint_filename)
             return model
 
         elif self.backend == "keras":
-            from zoo.zouwu.model.forecast.model import KerasBaseModel
+            from zoo.zouwu.model.forecast.model.base_keras_model import KerasBaseModel
             model = KerasBaseModel(**self.params)
             model.restore(checkpoint_filename)
             return model
