+++ pyzoo/zoo/zouwu/regression/xgb_predictor.py
@@ -17,7 +17,7 @@
 from zoo.automl.regression.base_predictor import BasePredictor
 from zoo.zouwu.feature.identity_transformer import IdentityTransformer
 
-from zoo.zouwu.model.forecast.model import XGBoost
+from zoo.zouwu.model.forecast.model.XGBoost import XGBoost
 
 
 class XgbPredictor(BasePredictor):
